pyproject.toml
[project]
name = "docx-navigator"
version = "0.1.0"
description = "Merge multiple Word (.docx) files with a clickable navigation menu"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-docx>=1.1.0",
    "docxcompose>=1.4.0",
    "lxml>=5.0.0",
    "click>=8.1.0",
]

[project.scripts]
docx-navigator = "docx_navigator.main:main"
